Refactor table utilities
